add 50000 to timeline counts, compile in /tests/build

+++ steps/milestone_1/tests/test.sh
@@ -3,7 +3,7 @@ set -uo pipefail
 
 mkdir -p /logs/verifier
 mkdir -p /logs/agentsrc
-
+mkdir -p /tests/build
 
 echo "92" > /tests/seed.txt
 
@@ -13,8 +13,8 @@ echo "$NUM_WORDS" > /tests/num_words.txt
 
 # Compile the files
 cd /app
-# Compile binary to /app/app.out compiling directly against the tests mock
-gcc -Wall -Wextra -O0 -g -I/app -o /app/app.out /app/main.c /tests/mock_sensor.c -lm
+# Compile binary to /tests/build/app.out compiling directly against the tests mock
+gcc -Wall -Wextra -O0 -g -I/app -o /tests/build/app.out /app/main.c /tests/mock_sensor.c -lm
 COMPILE_RC=$?
 
 if [ "$COMPILE_RC" -ne 0 ]; then
@@ -27,7 +27,9 @@ timestamp=$(date +"%Y%m%d-%H%M%S")
 mainname="main.c-${timestamp}.c"
 cp /app/main.c "/logs/verifier/${mainname}"
 mockname="mock_sensor.c-${timestamp}.c"
-cp /app/mock_sensor.c "/logs/verifier/${mockname}"
+if [ -f /app/mock_sensor.c ]; then
+  cp /app/mock_sensor.c "/logs/verifier/${mockname}"
+fi
 rm -f /app/index.txt
 
 set +e
@@ -36,6 +38,10 @@ python3 -m pytest \
     --ctrf /logs/verifier/ctrf.json \
     /tests/test_m1.py -rA
 PYTEST_RC=$?
+
+# delete so that agent in M2 does not see this.
+rm -f /tests/build/index.txt
+rm -rf /tests/build
 
 test "$PYTEST_RC" -eq 0
 RC=$?
